Added 'users/lookup' helper method and more tests.
First, the twitter spec requires that 'users/show' have a parameter
named `user_id`, not just `id`.

Second, added a new sugar method `#getUsersLookup` which maps to
`users/lookup`, as well as tests.

Third, added a new testcase for both `#getUsers()` and
`#getUsersLookup()` that ensures that an exception is thrown when
required parameters are not present.