Update MediaTrait.php
last version was expecting, for the video branch, a path to the video. Then it did file_get_contents, etc. To make both branches the same (image and video), we are now expectin parameter['media'] to be the contents in bytes of the file. That way, its the consumer responsibility to do file_get_contents($path) to fetch the contents of the video. 
It also updates the getMediaMime to read from a stream instead of a path.